+++ astro.config.mjs
@@ -8,5 +8,6 @@ export default defineConfig({
   devToolbar: {
     enabled: false
   },
+  output: 'server',
   integrations: [tailwind(), preact()]
 });