+++ content/messages.js
@@ -1,0 +1,167 @@
+export default {
+  "Luis Emidio | Homepage": {
+    "pt": "Luis Emidio | Página Inicial"
+  },
+  "Explore my personal website.": {
+    "pt": "Explore o meu site pessoal."
+  },
+  "Luis Emidio, website, technology, developer": {
+    "pt": "Luis Emidio, site, tecnologia, desenvolvedor"
+  },
+  "Let's connect!": {
+    "pt": "Vamos nos conectar!"
+  },
+  "That's all it is: information. Even a simulated experience or a dream; simultaneous reality and fantasy. Any way you look at it, all the information that a person accumulates in a lifetime is just a drop in the bucket.\" — Batou": {
+    "pt": "Isso é tudo: informação. Mesmo uma experiência simulada ou um sonho; realidade e fantasia simultâneas. De qualquer forma que você olhe para isso, toda a informação que uma pessoa acumula em uma vida é apenas uma gota no balde.\" — Batou"
+  },
+  "Full Stack Developer": {
+    "pt": "Desenvolvedor Full Stack"
+  },
+  "About me": {
+    "pt": "Sobre mim"
+  },
+  "Who I am": {
+    "pt": "Quem sou eu?"
+  },
+  "I'm a natural neural network navigating the digital cosmos, powered by curiosity and a relentless drive to explore the future. As a self-taught Full Stack Web Developer, I merge technology and creativity, weaving code into innovative solutions. My journey unfolds in the realm of cybernetic creativity, where every line of code is a brushstroke on the canvas of tomorrow.": {
+    "pt": "Sou uma rede neural natural navegando pelo cosmos digital, impulsionado pela curiosidade e um impulso implacável para explorar o futuro. Como um Desenvolvedor Full Stack autodidata, eu fundo tecnologia e criatividade, tecendo código em soluções inovadoras. Minha jornada se desdobra no reino da criatividade cibernética, onde cada linha de código é uma pincelada no quadro do amanhã."
+  },
+  "My mission": {
+    "pt": "Minha missão"
+  },
+  "My mission is deeply rooted in the observation that technology evolves exponentially and rapidly, far outpacing the pace of advancements in other contexts. I aim to explore and innovate in the intersection of technology and other realms, with a focus on understanding how technology can be seamlessly integrated into various aspects of life, from professional endeavors in tech development to hobbies. This mission involves fostering a holistic approach to embracing technology's potential, creating a convergence where slower aspects can catch up and align with the lightning-fast pace of technological progress.": {
+    "pt": "Minha missão está profundamente enraizada na observação de que a tecnologia evolui de forma exponencial e rápida, superando em muito o ritmo dos avanços em outros contextos. Meu objetivo é explorar e inovar na interseção da tecnologia e de outros domínios, com foco em compreender como a tecnologia pode ser integrada de forma perfeita em vários aspectos da vida, desde empreendimentos profissionais no desenvolvimento de tecnologia até hobbies. Esta missão envolve fomentar uma abordagem holística para abraçar o potencial da tecnologia, criando uma convergência onde os aspectos mais lentos podem alcançar e se alinhar com o ritmo veloz do progresso tecnológico."
+  },
+  "TL;DR": {
+    "pt": "Resumo"
+  },
+  "Embracing tech integration": {
+    "pt": "Abraçar a integração tecnológica"
+  },
+  "My skills": {
+    "pt": "Minhas habilidades"
+  },
+  "Skills, whether innate or acquired, are instrumental in fulfilling my mission to embrace and harness the power of technology for transformative endeavors.": {
+    "pt": "Habilidades, sejam inatas ou adquiridas, são instrumentais para cumprir minha missão de abraçar e aproveitar o poder da tecnologia para empreendimentos transformadores."
+  },
+  "I'm not one to often list my skills publicly, as they constantly evolve through continuous study and improvement. However, I'd like to highlight some innate rare skills that I leverage:": {
+    "pt": "Não costumo listar minhas habilidades publicamente, pois elas evoluem constantemente por meio de estudo e melhoria contínuos. No entanto, gostaria de destacar algumas habilidades raras inatas que eu utilizo:"
+  },
+  "Want to know more about me?": {
+    "pt": "Quer saber mais sobre mim?"
+  },
+  "I've curated a collection of my favorite content that resonate deeply with me. If you are curious, you can explore it here:": {
+    "pt": "Criei uma coleção do meu conteúdo favorito que ressoa profundamente comigo. Se você está curioso, pode explorá-lo aqui:"
+  },
+  "My YouTube playlists": {
+    "pt": "Minhas playlists do YouTube"
+  },
+  "Contact": {
+    "pt": "Contato"
+  },
+  "If you share similar interests or want to chat about exciting futuristic projects, don't hesitate to reach out. I would love to hear from you!": {
+    "pt": "Se você compartilha interesses semelhantes ou quer conversar sobre projetos futuristas empolgantes, não hesite em entrar em contato. Adoraria ouvir de você!"
+  },
+  "Let's be part of the future together!": {
+    "pt": "Vamos fazer parte do futuro juntos!"
+  },
+  "Something important?": {
+    "pt": "Algo importante?"
+  },
+  "Send me an email at": {
+    "pt": "Envie-me um e-mail para"
+  },
+  "Just want to chat?": {
+    "pt": "Só quer conversar?"
+  },
+  "Send me a message on Discord": {
+    "pt": "Envie-me uma mensagem no Discord"
+  },
+  "Portfolio": {
+    "pt": "Portfólio"
+  },
+  "I'm constantly striving to enhance my skills and learn new things. Feel free to check back in the future to see what I've been working on. Thank you for your interest in my work, and I look forward to sharing more with you soon!": {
+    "pt": "Estou constantemente buscando aprimorar minhas habilidades e aprender coisas novas. Sinta-se à vontade para verificar no futuro para ver no que tenho trabalhado. Obrigado pelo seu interesse no meu trabalho, e estou ansioso para compartilhar mais com você em breve!"
+  },
+  "Recent Repositories": {
+    "pt": "Repositórios Recentes"
+  },
+  "Click on the icons below to visit the respective source code or demo for more details.": {
+    "pt": "Clique nos ícones abaixo para visitar o código-fonte ou a demonstração respectivos para mais detalhes."
+  },
+  "Click on the icons to visit the respective source code or demo for more details.": {
+    "pt": "Clique nos ícones para visitar o código-fonte ou a demonstração respectivos para mais detalhes."
+  },
+  "Explore all my repositories": {
+    "pt": "Explore todos os meus repositórios"
+  },
+  "Access my GitHub Profile": {
+    "pt": "Acesse meu Perfil do GitHub"
+  },
+  "Currently, I'm in the process of building my portfolio. Meanwhile, you can track my progress and explore some of my work on": {
+    "pt": "Atualmente, estou no processo de construção do meu portfólio. Enquanto isso, você pode acompanhar meu progresso e explorar alguns dos meus trabalhos em"
+  },
+  "I learn things fast": {
+    "pt": "Eu aprendo rápido"
+  },
+  "I possess a keen ability to swiftly absorb new concepts and technologies, enabling me to stay ahead in a fast-paced tech environment.": {
+    "pt": "Possuo uma habilidade aguçada para absorver rapidamente novos conceitos e tecnologias, permitindo-me manter-me à frente em um ambiente de tecnologia acelerado."
+  },
+  "The new cativates me": {
+    "pt": "O novo me cativa"
+  },
+  "I am constantly captivated by new developments in technology, maintaining a perpetual curiosity that drives me to explore without succumbing to boredom.": {
+    "pt": "Sou constantemente cativado por novos desenvolvimentos na tecnologia, mantendo uma curiosidade perpétua que me impulsiona a explorar sem sucumbir ao tédio."
+  },
+  "I think computationally": {
+    "pt": "Eu penso de forma computacional"
+  },
+  "I excel at computational thinking, leveraging analytical skills and innovative approaches to solve complex tech problems.": {
+    "pt": "Excelho no pensamento computacional, utilizando habilidades analíticas e abordagens inovadoras para resolver problemas tecnológicos complexos."
+  },
+  "I have a creative mindset": {
+    "pt": "Eu tenho uma mentalidade criativa"
+  },
+  "I possess a creative mindset, constantly generating new ideas and thinking outside the box when tackling tech-related challenges.": {
+    "pt": "Possuo uma mentalidade criativa, gerando constantemente novas ideias e pensando fora da caixa ao enfrentar desafios relacionados à tecnologia."
+  },
+  "I am future-focused": {
+    "pt": "Eu estou focado no futuro"
+  },
+  "I prioritize potential over past knowledge, embracing a deep understanding of forthcoming technologies and focusing on what I can deeply continue to comprehend and improve even more in the future.": {
+    "pt": "Priorizo o potencial sobre o conhecimento passado, abraçando uma compreensão profunda das tecnologias futuras e concentrando-me no que posso continuar a compreender e aprimorar ainda mais no futuro."
+  },
+  "I like to go deep on topics": {
+    "pt": "Eu gosto de aprofundar nos temas"
+  },
+  "I have a passion for delving deep into topics, exploring intricate details and gaining a comprehensive understanding that goes beyond surface-level knowledge. This depth of exploration enables me to uncover valuable insights and innovative solutions.": {
+    "pt": "Tenho paixão por aprofundar em temas, explorando detalhes intrincados e adquirindo uma compreensão abrangente que vai além do conhecimento superficial. Essa profundidade de exploração me permite descobrir insights valiosos e soluções inovadoras."
+  },
+  "Discord copied successfully!": {
+    "pt": "Discord copiado com sucesso!"
+  },
+  "Error copying Discord!": {
+    "pt": "Erro ao copiar o Discord!"
+  },
+  "{year} {separator} Generated by {ai} + {me}": {
+    "pt": "{year} {separator} Gerado por {ai} + {me}"
+  },
+  "AI": {
+    "pt": "IA"
+  },
+  "There was an error loading projects from GitHub. 😭": {
+    "pt": "Ocorreu um erro ao carregar os projetos do GitHub. 😭"
+  },
+  "Home": {
+    "pt": "Início"
+  },
+  "About": {
+    "pt": "Sobre"
+  },
+  "English": {
+    "pt": "Inglês"
+  },
+  "Portuguese": {
+    "pt": "Português"
+  },
+}