+++ CHANGELOG.md
@@ -5,3 +5,13 @@ description: All notable changes to this project will be documented in this file
 ---
 
 ## [Unreleased]
+
+### Added
+
+- Added hero with my name and role
+- Added "about me" section, including "who I am", mission, skills and content link
+- Added portfolio section with recent repos
+- Added contact section with links
+- Added footer with page links
+- Added light theme
+- Added Brazilian Portuguese translation